Added content to the readme for Rust Eventhubs; added documentation for the Consumer module
